Added code files.
Added code files to repository, These files implement the portion of the project which captures posts from Gab and Twitter in real time based on their hashtags. Then using Python I perform sentiment analysis on the content of the posts and determine their sentiment score. The posts themselves and their sentiment score are then uploaded to a SQL server for storage.